Security fixes: OOM protection, u128 truncation, RPC bind localhost, div-by-zero guard
- P1-T3: Add bounds checking in deserialize_proofs (merkle.rs) to prevent OOM
  attacks via crafted data with huge num_proofs (>10,000) or num_siblings (>64)
- P1-T6: Change TransactionMessage.value from u64 to u128 (protocol.rs) and
  remove truncating casts in node/lib.rs to prevent value truncation
- P1-T7: Bind RPC to 127.0.0.1:8545 by default in testnet() and mainnet()
  configs instead of 0.0.0.0 to prevent unintended external exposure
- P1-T9: Redis authentication already addressed (requirepass in docker-compose)
- Add division-by-zero guard in generate_samples (consensus/lib.rs) for
  output_size == 0

+++ core/core/network/src/protocol.rs
@@ -37,7 +37,7 @@ pub struct TransactionMessage {
     pub hash: String,
     pub from: String,
     pub to: String,
-    pub value: u64,
+    pub value: u128,
     pub data: Vec<u8>,
     pub nonce: u64,
     pub signature: Vec<u8>,
